+++ Example/AndesUI/Messages/Views/MessageViewController.swift
@@ -30,11 +30,16 @@ class MessageViewController: UIViewController, MessageView {
     @IBOutlet weak var dismissibleSwitch: UISwitch!
     @IBOutlet weak var errorLabel: UILabel!
     @IBOutlet weak var showAgainBtn: AndesButton!
+    @IBOutlet weak var showBulletsSwitch: UISwitch!
 
     var typePicker: UIPickerView = UIPickerView()
     var statePicker: UIPickerView = UIPickerView()
     var type: AndesMessageType = .neutral
     var hierarchy: AndesMessageHierarchy = .loud
+    var availableBullets: [AndesBullet] = []
+    var showBullets: Bool {
+        self.showBulletsSwitch.isOn
+    }
 
     fileprivate func setupButtons() {
         configToggleButton.text = "message.button.changeConfig".localized
@@ -84,6 +89,31 @@ class MessageViewController: UIViewController, MessageView {
         messageView.setSecondaryAction("Secondary", handler: {[unowned self] _ in self.didPressButton()})
     }
 
+    private func bulletSetup() {
+        let secondBulletLinks = [
+            AndesBodyLink(startIndex: 0, endIndex: 26),
+            AndesBodyLink(startIndex: 38, endIndex: 44)
+        ]
+
+        let links = AndesBodyLinks(links: []) { _ in }
+
+        let bodyLinksForSecondBullet = AndesBodyLinks(links: secondBulletLinks) { [unowned self] index in
+            self.didBulletPressBodyLink(index)
+        }
+
+        let bullet1 = AndesBullet(text: "Bullet 1 example.", bodyLink: links)
+        let bullet2 = AndesBullet(text: "Bullet 2 Multiline example with simple dummy text of the printing and tysetting industry. Lorem impsum.", bodyLink: bodyLinksForSecondBullet)
+        let bullet3 = AndesBullet(text: "Bullet 3 example.", bodyLink: links)
+
+        let bullets = [
+            bullet1,
+            bullet2,
+            bullet3
+        ]
+
+        self.availableBullets = bullets
+    }
+
     override func viewDidLoad() {
         super.viewDidLoad()
         setupButtons()
@@ -91,6 +121,7 @@ class MessageViewController: UIViewController, MessageView {
         createPickerViews()
 
         setupBaseMessage()
+        bulletSetup()
     }
 
     func didPressButton() {
@@ -100,7 +131,15 @@ class MessageViewController: UIViewController, MessageView {
     }
 
     func didPressBodyLink(_ index: Int) {
-        let alert = UIAlertController(title: "message.actions.pressedMsg".localized, message: "Body link pressed at position: \(index)", preferredStyle: .alert)
+        self.showBodyLinkMessageAlert(index)
+    }
+
+    func didBulletPressBodyLink(_ index: Int) {
+        self.showBodyLinkMessageAlert(linkType: "Bullet body link", index)
+    }
+
+    func showBodyLinkMessageAlert(linkType: String = "Body link", _ index: Int) {
+        let alert = UIAlertController(title: "message.actions.pressedMsg".localized, message: "\(linkType) pressed at position: \(index)", preferredStyle: .alert)
         alert.addAction(UIAlertAction(title: "OK", style: .default, handler: nil))
         self.present(alert, animated: true, completion: nil)
     }
@@ -145,7 +184,7 @@ class MessageViewController: UIViewController, MessageView {
     }
 
     func validate(body: String, primary: String, secondary: String, link: String) -> String? {
-        guard !body.isEmpty else {
+        guard !body.isEmpty || showBullets else {
             return "message.error.emptyBody".localized
         }
         guard secondary.isEmpty || !primary.isEmpty else {
@@ -169,6 +208,7 @@ class MessageViewController: UIViewController, MessageView {
         let primary = primaryActionTextField.text!
         let secondary = secondaryActionTextField.text!
         let link = linkActionTextField.text!
+
         if let err = validate(body: body, primary: primary, secondary: secondary, link: link) {
             errorLabel.text = err
             errorLabel.isHidden = false
@@ -183,6 +223,8 @@ class MessageViewController: UIViewController, MessageView {
         messageView.isDismissable = dismiss
         messageView.type = type
         messageView.hierarchy = hierarchy
+        messageView.bullets = showBullets ? self.availableBullets : []
+
         messageView.setPrimaryAction(primary, handler: {[unowned self] _ in self.didPressButton()})
         messageView.setSecondaryAction(secondary, handler: {[unowned self] _ in self.didPressButton()})
         messageView.setLinkAction(link, handler: {[unowned self] _ in self.didPressButton()})