+++ CHANGELOG.md
@@ -1,4 +1,17 @@
-# Unpublished
+# 3.24.0
+### 🚀 Features
+- New component Tooltip | [Juan Vasquez](https://github.com/javff)
+- Added Bullets for Andes Message | Authors [Martin Becerra](https://github.com/mabecerra-MeLi)
+
+# 3.23.2
+### 🛠 Bug fixes
+- Optional color var | Authors [joalonspint](https://github.com/joalonsopint)
+
+# 3.23.0
+### 🛠 Bug fixes
+- Fix color name and color in colorset | Authors [joalonspint](https://github.com/joalonsopint)
+
+# 3.22.0
 ### ⚙️ Other
 - Home router refactored | Authors [jbrunoAtMeli] (https://github.com/jbrunoAtMeli)
 - Lib & testapp updated config: | Authors [jbrunoAtMeli] (https://github.com/jbrunoAtMeli)